Now can handle longs and change to recursive with fix to actually use memorization table
Earlier version always had input as 1 when it went to set the
memorization value.

+++ src/main.h
@@ -31,4 +31,4 @@
 
 void runserver();
 void *handle_client(void *arg);
-int three_a_one(int input);
+int three_a_one(long input);
